Force fullpage_redirect views to use raw values, not escaped (issue #68 potential fix), use bigInt for charge IDs to fix issue #65

+++ src/ShopifyApp/resources/database/migrations/2018_06_03_185902_create_charges_table.php
@@ -18,7 +18,7 @@ class CreateChargesTable extends Migration
             $table->increments('id');
 
             // Filled in when the charge is created, provided by shopify, unique makes it indexed
-            $table->integer('charge_id')->unique();
+            $table->bigInteger('charge_id')->unique();
 
             // Test mode or real
             $table->boolean('test');
